:bug: | Corrigir erros no "+mencionar"

+++ src/main/java/com/mrpowergamerbr/loritta/commands/vanilla/misc/QuoteCommand.kt
@@ -3,12 +3,17 @@ package com.mrpowergamerbr.loritta.commands.vanilla.misc
 import com.mrpowergamerbr.loritta.commands.CommandBase
 import com.mrpowergamerbr.loritta.commands.CommandCategory
 import com.mrpowergamerbr.loritta.commands.CommandContext
+import com.mrpowergamerbr.loritta.utils.LorittaUtils
 import com.mrpowergamerbr.loritta.utils.getOrCreateWebhook
+import com.mrpowergamerbr.loritta.utils.isValidSnowflake
+import com.mrpowergamerbr.loritta.utils.locale.BaseLocale
 import com.mrpowergamerbr.temmiewebhook.DiscordEmbed
 import com.mrpowergamerbr.temmiewebhook.DiscordMessage
 import com.mrpowergamerbr.temmiewebhook.embed.AuthorEmbed
 import com.mrpowergamerbr.temmiewebhook.embed.FooterEmbed
 import net.dv8tion.jda.core.Permission
+import net.dv8tion.jda.core.entities.Message
+import net.dv8tion.jda.core.exceptions.ErrorResponseException
 import java.util.*
 
 class QuoteCommand : CommandBase() {
@@ -16,8 +21,8 @@ class QuoteCommand : CommandBase() {
 		return "mencionar"
 	}
 
-	override fun getDescription(): String {
-		return "Menciona uma mensagem, dica: para copiar o ID da mensagem, ative o modo de desenvolvedor do Discord e clique com botão direito em uma mensagem!"
+	override fun getDescription(locale: BaseLocale): String {
+		return locale.get("MENCIONAR_DESCRIPTION")
 	}
 
 	override fun hasCommandFeedback(): Boolean {
@@ -36,39 +41,48 @@ class QuoteCommand : CommandBase() {
 		if (context.args.size >= 1) {
 			val temmie = getOrCreateWebhook(context.event.textChannel, "Quote Webhook")
 
-			val msg = context.event.textChannel.getMessageById(context.args[0]).complete() ?: return
+			if (context.args[0].isValidSnowflake()) {
+				var msg: Message? = null
+				try {
+					msg = context.event.textChannel.getMessageById(context.args[0]).complete()
+				} catch (e: ErrorResponseException) {
+				}
+				if (msg != null) {
+					if (context.guild.selfMember.hasPermission(Permission.MESSAGE_MANAGE)) {
+						context.message.delete().complete() // ok, vamos deletar a msg original
+					}
 
-			if (context.guild.selfMember.hasPermission(Permission.MESSAGE_MANAGE)) {
-				context.message.delete().complete() // ok, vamos deletar a msg original
+					var content = msg.author.asMention + " " + context.event.message.rawContent.replace(context.config.commandPrefix + "mencionar " + context.args[0], "").trim { it <= ' ' }
+					content = content.replace("@here", "")
+					content = content.replace("@everyone", "")
+
+					val embed = DiscordEmbed
+							.builder()
+							.author(AuthorEmbed(msg.author.name + " ${context.locale.get("MENCIONAR_SAID")}...", null, msg.author.effectiveAvatarUrl, null))
+							.color(123)
+							.description(msg.rawContent)
+							// .title("Wow!")
+							.footer(FooterEmbed("em #" + context.message.textChannel.name + if (context.guild.selfMember.hasPermission(Permission.MESSAGE_MANAGE)) "" else " | Não tenho permissão para deletar mensagens!", null, null))
+							.build()
+
+					val dm = DiscordMessage
+							.builder()
+							.avatarUrl(context.message.author.effectiveAvatarUrl)
+							.username(context.message.author.name)
+							.content(content)
+							.embed(embed)
+							.build()
+
+
+					dm.embeds = Arrays.asList(embed)
+
+					temmie!!.sendMessage(dm)
+				} else {
+					context.sendMessage(LorittaUtils.ERROR + " **|** ${context.getAsMention(true)}" + context.locale.get("MENCIONAR_UNKNOWN_MESSAGE"))
+				}
+			} else {
+				context.sendMessage(LorittaUtils.ERROR + " **|** ${context.getAsMention(true)}" + context.locale.get("MENCIONAR_NOT_VALID_SNOWFLAKE", context.args[0]))
 			}
-
-			var content = msg.author.asMention + " " + context.event.message.rawContent.replace(context.config.commandPrefix + "mencionar " + context.args[0], "").trim { it <= ' ' }
-			content = content.replace("@here", "")
-			content = content.replace("@everyone", "")
-
-			val embed = DiscordEmbed
-					.builder()
-					.author(AuthorEmbed(msg.author.name + " disse...", null, msg.author.effectiveAvatarUrl, null))
-					.color(123)
-					.description(msg.rawContent)
-					// .title("Wow!")
-					.footer(FooterEmbed("em #" + context.message.textChannel.name + if (context.guild.selfMember.hasPermission(Permission.MESSAGE_MANAGE)) "" else " | Não tenho permissão para deletar mensagens!", null, null))
-					.build()
-
-			val dm = DiscordMessage
-					.builder()
-					.avatarUrl(context.message.author.effectiveAvatarUrl)
-					.username(context.message.author.name)
-					.content(content)
-					.embed(embed)
-					.build()
-
-
-			dm.embeds = Arrays.asList(embed)
-
-			temmie!!.sendMessage(dm)
-
-			println("Enviado!")
 		} else {
 			context.explain()
 		}
